Move subnet IP allocation to its own file
